♻️ move routes out of portalcontainer (#2102)
