+++ src/packages/@ncigdc/components/PortalContainer.js
@@ -2,38 +2,20 @@
 
 import React from 'react';
 import { connect } from 'react-redux';
-import { Route, Switch } from 'react-router-dom';
 import Cookies from 'js-cookie';
 import { compose, lifecycle } from 'recompose';
-
-import ProjectRoute from '@ncigdc/routes/ProjectRoute';
-import FileRoute from '@ncigdc/routes/FileRoute';
-import CaseRoute from '@ncigdc/routes/CaseRoute';
-import AnnotationRoute from '@ncigdc/routes/AnnotationRoute';
-
-import ComponentsRoute from '@ncigdc/routes/ComponentsRoute';
-import GeneRoute from '@ncigdc/routes/GeneRoute';
-import AnalysisRoute from '@ncigdc/routes/AnalysisRoute';
-import SSMRoute from '@ncigdc/routes/SSMRoute';
-import ManageSetsRoute from '@ncigdc/routes/ManageSetsRoute';
-import SmartSearchRoute from '@ncigdc/routes/SmartSearchRoute';
-
-import Head from '@ncigdc/components/Head';
 import Header from '@ncigdc/components/Header';
 import Footer from '@ncigdc/components/Footer';
 import NotificationContainer from '@ncigdc/components/NotificationContainer';
 import RelayLoadingContainer from '@ncigdc/components/RelayLoadingContainer';
 import ProgressContainer from '@ncigdc/components/ProgressContainer';
 import ModalContainer from '@ncigdc/components/ModalContainer';
-import NotFound from '@ncigdc/components/NotFound';
-
+import Routes from '@ncigdc/routes';
 import withRouter from '@ncigdc/utils/withRouter';
 import { GlobalTooltip } from '@ncigdc/uikit/Tooltip';
-
 import styled from '@ncigdc/theme/styled';
 import { setModal } from '@ncigdc/dux/modal';
 import FirstTimeModal from '@ncigdc/components/Modals/FirstTimeModal';
-import LoadableWithLoading from '@ncigdc/components/LoadableWithLoading';
 
 const SkipLink = styled.a({
   position: 'absolute',
@@ -43,30 +25,6 @@ const SkipLink = styled.a({
 });
 
 const FIRST_TIME_KEY = 'NCI-Warning';
-
-const HomeRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/HomeRoute'),
-});
-
-const CartRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/CartRoute'),
-});
-
-const RepositoryRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/RepositoryRoute'),
-});
-
-const ExploreRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/ExploreRoute'),
-});
-
-const ProjectsRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/ProjectsRoute'),
-});
-
-const AnnotationsRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/AnnotationsRoute'),
-});
 
 const enhance = compose(
   withRouter,
@@ -116,31 +74,9 @@ const PortalContainer = ({
         transition: 'padding 0.25s ease',
       }}
     >
-      <Route
-        children={p => <Head title={p.location.pathname.split('/')[1]} />}
-      />
-      <Switch>
-        <Route exact path="/" component={HomeRoute} />
-        <Route exact path="/cart" component={CartRoute} />
-        <Route exact path="/repository" component={RepositoryRoute} />
-        <Route exact path="/exploration" component={ExploreRoute} />
-        <Route exact path="/projects" component={ProjectsRoute} />
-        <Route exact path="/annotations" component={AnnotationsRoute} />
-        <Route exact path="/query" component={SmartSearchRoute} />
-        {ProjectRoute}
-        <Route path="/files/:id" component={FileRoute} />
-        {CaseRoute}
-        {AnnotationRoute}
-        {GeneRoute}
-        {ManageSetsRoute}
-        {AnalysisRoute}
-        {SSMRoute}
-        <Route path="/components/:component" component={ComponentsRoute} />
-        <Route component={NotFound} />
-      </Switch>
+      <Routes />
     </div>
     <Footer />
-
     <RelayLoadingContainer />
     <NotificationContainer />
     <ModalContainer />
